+++ cluster/proposer.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"log"
 	"math/rand"
-	"reflect"
 )
 
 var (
@@ -14,34 +13,20 @@ var (
 )
 
 type (
-	// Proposal represents an operation that is sent to the cluster.
-	// Users are expected to define their own structs that embed ProposalBase
-	// to satisfy this interface.
-	Proposal interface {
-		// ProposalID returns a unique ID for this proposal.
-		ProposalID() uint64
-	}
-
-	// Response is returned from a HandlerFunc.
-	Response interface {
-		Error() string
-	}
-
-	// HandlerFunc is a function that handles committing a Proposal once it has been
-	// accepted by the cluster.
-	HandlerFunc func(p Proposal) Response
+	ProposalType uint32
+	// ProposalFunc handles committing a proposal once it has been accepted by the cluster.
+	ProposalFunc func(data []byte) (resp any, err error)
 
 	// Proposer encapsulates making proposals to a cluster,
 	// and handling those proposals once they are accepted.
 	Proposer interface {
-		// Handle registers a function that will handle accepted proposals.
-		// Argument p should be a pointer to a concrete type that implements Proposal.
-		// Handle panics if it is called more than once for a single concrete type.
-		Handle(p Proposal, f HandlerFunc)
+		// Handle registers a function that will handle accepted proposals for the given type.
+		// Handle panics if it is called more than once for a single type.
+		Handle(t ProposalType, f ProposalFunc)
 		// Propose makes a proposal to the cluster.
 		// Propose panics if a HandlerFunc has not been registered
-		// for the given proposal type using Handle.
-		Propose(p Proposal) Response
+		//  for the given proposal type using Handle.
+		Propose(t ProposalType, data []byte) (resp any, err error)
 	}
 )
 
@@ -69,28 +54,26 @@ func (r *ResponseBase) Error() string {
 // NewFakeProposer returns a new FakeProposer.
 func NewFakeProposer() *FakeProposer {
 	return &FakeProposer{
-		handlerFuncs: make(map[reflect.Type]HandlerFunc),
+		handlerFuncs: make(map[ProposalType]ProposalFunc),
 	}
 }
 
 // FakeProposer handles proposals without actually submitting them to a cluster.
 type FakeProposer struct {
-	handlerFuncs map[reflect.Type]HandlerFunc
+	handlerFuncs map[ProposalType]ProposalFunc
 
 	// Proposals is incremented for every submitted proposal.
 	Proposals int64
 }
 
-func (p *FakeProposer) Handle(prop Proposal, f HandlerFunc) {
-	t := reflect.TypeOf(prop)
+func (p *FakeProposer) Handle(t ProposalType, f ProposalFunc) {
 	if _, ok := p.handlerFuncs[t]; ok {
 		log.Fatalf("proposal type already registered: %s", t)
 	}
 	p.handlerFuncs[t] = f
 }
 
-func (p *FakeProposer) Propose(prop Proposal) Response {
-	t := reflect.TypeOf(prop)
+func (p *FakeProposer) Propose(t ProposalType, data []byte) (resp any, err error) {
 	f, ok := p.handlerFuncs[t]
 	if !ok {
 		panic(fmt.Sprintf("unknown proposal type received: %s", t))
@@ -98,5 +81,5 @@ func (p *FakeProposer) Propose(prop Proposal) Response {
 
 	p.Proposals++
 
-	return f(prop)
+	return f(data)
 }
